Change datatype of mb_mode_info structure members
Changed datatype of 'comp_group_idx', 'compound_idx', 'num_proj_ref'
and 'overlappable_neighbors[]' from int to uint8_t.

Observed memory footprint reduction with no impact on
encoder/decoder performance.

Stream                   Preset     Encoder   Decoder
BasketballDrill_832x480     1        ~0.3%       ~2.3%
parkrun_720p50              3        ~1%         ~2.7%

+++ av1/decoder/decodemv.c
@@ -1412,14 +1412,14 @@ static void read_inter_block_mode_info(AV1Decoder *const pbi,
 
     if (masked_compound_used) {
       const int ctx_comp_group_idx = get_comp_group_idx_context(xd);
-      mbmi->comp_group_idx = aom_read_symbol(
+      mbmi->comp_group_idx = (uint8_t)aom_read_symbol(
           r, ec_ctx->comp_group_idx_cdf[ctx_comp_group_idx], 2, ACCT_STR);
     }
 
     if (mbmi->comp_group_idx == 0) {
       if (cm->seq_params.order_hint_info.enable_dist_wtd_comp) {
         const int comp_index_ctx = get_comp_index_context(cm, xd);
-        mbmi->compound_idx = aom_read_symbol(
+        mbmi->compound_idx = (uint8_t)aom_read_symbol(
             r, ec_ctx->compound_index_cdf[comp_index_ctx], 2, ACCT_STR);
         mbmi->interinter_comp.type =
             mbmi->compound_idx ? COMPOUND_AVERAGE : COMPOUND_DISTWTD;
